sitecore-ai-automation/adaptive-agent-orchestrator/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "adaptive-agent-orchestrator"
version = "1.0.0"
description = "Reliable orchestrator for Sitecore AI automation — LangGraph + Claude Code subscription"
requires-python = ">=3.11"
dependencies = [
    "langchain-claude-code>=0.1.0",
    "langgraph>=0.4.0",
    "langgraph-checkpoint>=2.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "langgraph-cli>=0.1.0",
]

[project.scripts]
orch = "cli:main"
